Fix CI: gate slow corpus fixtures in dask backends + fold #2449 test (#2440)
Two follow-ups after the consolidation surfaced fresh issues:

1. CI fast-lane failure (ImportError: lerc) in
   ``test_dask_numpy_parity[compression_lerc_float32]``. The original
   top-level ``test_golden_corpus_dask_numpy_1930.py`` was never
   reached by the ``pytest xrspatial/geotiff/tests/golden_corpus/``
   workflow target, so a latent bug in its ``_build_param`` (no call
   to ``fast_slow_marks_for``) was invisible. The Sub-PR B relocation
   moved the file into ``golden_corpus/`` per the issue spec, and the
   fast-lane CI lane started running it without the ``slow`` filter
   applying. Wire ``fast_slow_marks_for`` into the three dask /
   GPU backend modules (``test_dask_numpy``, ``test_dask_gpu``,
   ``test_gpu``) so the manifest's tagged-slow ``compression_*``
   fixtures (including the optional-dep ``lerc`` cell) drop out of
   ``pytest -m "not slow"`` the same way they do in the eager
   backend module. Nightly / push-to-main lanes still exercise the
   full set.

2. The merge from origin/main pulled in
   ``test_file_source_context_2449.py`` (PR #2479's coverage for the
   new ``_FileSource`` context-manager protocol). That landed at the
   top level with a numeric suffix, pushing the Definition-of-Done
   counts off-target (top-level=6, suffixed=1). Folded into
   ``read/test_basic.py`` alongside the other input-source variants
   so the counts return to 5 top-level / 0 suffixed.

Verified locally:
- ``pytest xrspatial/geotiff/tests/golden_corpus/ -m "not slow" -q``
  green with 30 deselected (the slow compression cells, including
  ``lerc``).
- ``pytest xrspatial/geotiff/tests/golden_corpus/ -q`` green (full
  lane: 339 passed, 25 skipped).
- ``pytest xrspatial/geotiff/tests/ -q`` green (5941 passed, 81
  skipped, 2 xfailed).
